refactor(codegen): extract std::process::* into stdlib/process.rs (Round 1.1)
10 lower fns + 2 helpers (load_i32_to_i64, store_struct_field) lifted
into stdlib/process.rs as the pub(crate) ProcessStdlib<'ctx> trait,
impl'd on Cx<'ctx, 'p>. Dispatch sites in codegen.rs (statement +
expression + fallible match arms) unchanged — trait import at the top
of codegen.rs resolves self.lower_std_process_*(...) calls.

Surface bumps to pub(crate):
- Cx fields: user_types, current_fn
- struct FallibleCallResult<'ctx>
- struct TypeInfo<'ctx> + struct_ty, fields
- Cx methods: unpack_view_if_needed, complete_io_fallible_call,
  alloca_in_entry, arena_alloc, emit_set_caller_arena

codegen.rs: 45459 -> 44690 (-769 lines).
